+++ orm/src/main/java/org/hibernate/tool/hbm2x/DocExporter.java
@@ -18,6 +18,7 @@ import org.hibernate.tool.hbm2x.doc.DocFile;
 import org.hibernate.tool.hbm2x.doc.DocFileManager;
 import org.hibernate.tool.hbm2x.doc.DocHelper;
 import org.hibernate.tool.hbm2x.pojo.POJOClass;
+import org.hibernate.tool.internal.export.AbstractExporter;
 
 /**
  * Exporter implementation that creates Hibernate Documentation.
