+++ orm/src/main/java/org/hibernate/tool/hbm2x/Hbm2DDLExporter.java
@@ -23,6 +23,7 @@ import java.util.Iterator;
 import org.hibernate.boot.Metadata;
 import org.hibernate.tool.hbm2ddl.SchemaExport;
 import org.hibernate.tool.hbm2ddl.SchemaExport.Action;
+import org.hibernate.tool.internal.export.AbstractExporter;
 import org.hibernate.tool.hbm2ddl.SchemaUpdate;
 import org.hibernate.tool.schema.TargetType;
