+++ orm/src/main/java/org/hibernate/tool/hbm2x/QueryExporter.java
@@ -11,6 +11,7 @@ import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
+import org.hibernate.tool.internal.export.AbstractExporter;
 
 /** 
  * exporter for query execution.
